shop and killfeed

+++ RagnarokBotWeb/Infrastructure/Repositories/Interfaces/IRepository.cs
@@ -5,7 +5,7 @@ namespace RagnarokBotWeb.Infrastructure.Repositories.Interfaces
     public interface IRepository<T> where T : class
     {
         Task<IEnumerable<T>> GetAllAsync();
-        Task<T> GetByIdAsync(int id);
+        Task<T> GetByIdAsync(long id);
         Task<IEnumerable<T>> FindAsync(Expression<Func<T, bool>> predicate);
         Task<T> FindOneAsync(Expression<Func<T, bool>> predicate);
         Task AddAsync(T entity);
